feat: 지도와 리스트를 통합하고, 현 위치를 기반으로 화면이 표시되게끔 바꿈.
