+++ src/app/detail/[id]/page.tsx
@@ -1,0 +1,6 @@
+const detailpage = () => {
+  return (
+    <div>detail</div>
+  )
+}
+export default detailpage
