+++ src/app/detail/page.tsx
@@ -1,0 +1,6 @@
+const paged = () => {
+  return (
+    <div>paged</div>
+  )
+}
+export default paged
