+++ src/types/Map.ts
@@ -1,0 +1,14 @@
+type coords = {
+  latitude: number;
+  longitude: number;
+};
+
+type latlng = {
+  lat: number;
+  lng: number;
+};
+
+type goCampingData = {
+  title: string;
+  latlng: latlng;
+};
